Extract version utils into separate file
